feat: support bfloat16 beta input with automatic cast
- chunk.py: cast bf16 beta to float32 before passing to kernels;
  backward gradient is cast back via db.to(beta)
- hopper_fused_fwd.py: cast beta to float32 before C++ kernel call
  (ElementBeta = float hardcoded in SM90 mainloop)
- blackwell_fused_fwd.py: cast beta to bfloat16 before CuteDSL kernel
  (io_dtype = BFloat16 in KDAChunkwise)
- README.md / USAGE.md: document that beta accepts both float32 and bfloat16
- tests: add bf16 beta test cases (B=2, T=512) to test_kda.py and
  test_kda_compare_fla.py

+++ cula/kda/hopper_fused_fwd.py
@@ -94,7 +94,7 @@ class HopperChunkKDAFunction(torch.autograd.Function):
         k = k.reshape(packed_seq, num_heads, head_dim).contiguous()
         v = v.reshape(packed_seq, num_heads, head_dim).contiguous()
         g = g.reshape(packed_seq, num_heads, head_dim).contiguous()
-        beta = beta.reshape(packed_seq, num_heads).contiguous()
+        beta = beta.reshape(packed_seq, num_heads).to(torch.float32).contiguous()
 
         # workspace buffer for TMA Store O tensormap
         sm_count = get_device_sm_count(q.device)
